refactor: rename com.mytaxi to com.freenow to prevent overriding by transitive dependencies
