+++ src/main/java/com/freenow/apis/phrase/api/format/Format.java
@@ -1,0 +1,12 @@
+package com.freenow.apis.phrase.api.format;
+
+import java.util.List;
+import org.apache.http.NameValuePair;
+
+public interface Format
+{
+
+    String getName();
+
+    List<NameValuePair> getOptions();
+}
