+++ src/main/java/com/freenow/apis/phrase/exception/PhraseAppApiException.java
@@ -1,0 +1,18 @@
+package com.freenow.apis.phrase.exception;
+
+public class PhraseAppApiException extends RuntimeException
+{
+    static final long serialVersionUID = -7034807190745766939L;
+
+
+    public PhraseAppApiException(String message)
+    {
+        super(message);
+    }
+
+
+    public PhraseAppApiException(String message, Throwable throwable)
+    {
+        super(message, throwable);
+    }
+}
