--- src/main/java/com/mytaxi/apis/phrase/api/format/Format.java
@@ -1,12 +1,0 @@
-package com.mytaxi.apis.phrase.api.format;
-
-import java.util.List;
-import org.apache.http.NameValuePair;
-
-public interface Format
-{
-
-    String getName();
-
-    List<NameValuePair> getOptions();
-}
